style: Extensions moved from DetailedView to standalone file
